MDEV-12547: InnoDB FULLTEXT index has too strict innodb_ft_result_cache_limit max limit
- Backported the MYSQL_SYSVAR_SIZE_T to 10.0
- The parameter innodb_ft_result_cache_limit was only 32 bits wide
also on 64-bit systems. Make it size_t, so that it will be 64 bits
on 64-bit systems.
- Added a test case that show how innodb_ft_result_cache_limit variables
behaves in 32bit and 64 bit system.

+++ storage/innobase/fts/fts0fts.cc
@@ -67,7 +67,7 @@ UNIV_INTERN ulong      fts_max_total_cache_size;
 
 /** This is FTS result cache limit for each query and would be
 a configurable variable */
-UNIV_INTERN ulong	fts_result_cache_limit;
+UNIV_INTERN size_t	fts_result_cache_limit;
 
 /** Variable specifying the maximum FTS max token size */
 UNIV_INTERN ulong	fts_max_token_size;
@@ -4308,7 +4308,7 @@ fts_sync_begin(
 	if (fts_enable_diag_print) {
 		ib_logf(IB_LOG_LEVEL_INFO,
 			"FTS SYNC for table %s, deleted count: %ld size: "
-			"%lu bytes",
+			"%zu bytes",
 			sync->table->name,
 			ib_vector_size(cache->deleted_doc_ids),
 			cache->total_size);
